+++ language/it/info_acp_portal.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * @package bbGuild Portal
+ * @copyright (c) 2026 avathar.be
+ * @license GNU General Public License, version 2 (GPL-2.0-only)
+ */
+
+if (!defined('IN_PHPBB'))
+{
+	exit;
+}
+
+if (empty($lang) || !is_array($lang))
+{
+	$lang = [];
+}
+
+$lang = array_merge($lang, [
+	'ACP_BBGUILD_PORTAL' => 'Portale',
+]);